Upgrade Validation version to 2.2
This commit also changes how we generate the interface we use as a mixin
PHP DocBlock, so it's more compatible with the rest of the codebase.

Because those interfaces are enormous, and the script ensures we fix all
the coding standards before putting them into place, I ignored them when
running PHP_CodeSniffer.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>