Make exception messages less verbose
Although the current messages can give the user much information, it
becomes unpractical and hard to understand what fails when making an
assertion.

I wasn't using this library when I made those customizations on the
exception messages. Now, getting to know more real-life use cases, like
validating a collection of objects, it has become more evident that the
messages are not legible as they could be.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>