Allow creation of assertions in any order
In the current design, one can only create assertions in specific orders
because each Creator passes on the creation to the following Creator
when they can't create an assertion.

To fix the issue, I had to abandon the concept of Assertors, which was
already bothering me. Instead, we're using custom Validation rules, and
with that approach, we can create nested rules with what we used to call
"Assertors."

Unfortunately, this was only possible with a gigantic commit because I
had to change the design of existing classes. Of course, I could have
created smaller commits, but they wouldn't be atomic.

I made changes even bigger by changing PHPStan to the max level, but
that makes sense since I'm starting a new design for the library.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>